nstore: group tuple items
This is easier to the mind to have:

  (prefix, permutation-index, (uid, key, value))

Than:

  (prefix, premutation-index, uid, key, value)

The tuple items are explicitly grouped in the key.
The only drawback is that packing above, will produce
a trailing null byte.  That is why, in  NStore.FROM it
is stripped with:

  tr.get_range_startswith(pack(prefix)[:-1]

It goes against the idea that: "flat is better than nested".
With the new approach, it is easier to read the content of
the database without having the play with slices.