nstore: support VersionStamp, vnstore: add it.
Instead of a general multiple branch support, like git does,
the current implementation maintains a list of addition, and
deletion, along the existing tuples that are possibly dead.

In other words, it is a single branch history, and a list of
patches.

When a change is applied, the history is not re-written, and
tuples are not copied, instead it will swap a `None` with a
VersionStamp.  VersionStamp with the help of transactions will
properly serialize history.

That allows to "merge" multiple changes concurrently,
if/when  foundationdb transactions are properly used.