README++
remove the git-like feature, indeed the last commit only support a
single branch history with possibly many patches that can be applied
to the history but are not part of a DAG.

Change-request, will be applied in-place, and required minimal change
to do so.

One drawback that must be handled later, is that the application of
change can lead to the same high level (uid, key, value, license)
appear multiple teams in vnstore.FROM and vnstore.where calls.

That will be handled by the application mechanic and the user (!)  to
avoid duplicate result.  I have no proof of that bug, however the
absence of proof is not the proof of absence, hence it is a possibily.

ref: #10