Do not mention git on the index page.
To avoid confusion, I prefer to remove the comparison with git in that
context.

In fact, git is used to store data and open data in several projects.
In that sense, copernic aims to replace git.  But it is not a
replacement for git/mercurial/fossil to track source code.